[c12640] Fix sending player cooldowns after map change/login and Correct logic in SpellMgr::IsSpellProcEventCanTriggeredBy

+++ src/game/Player.cpp
@@ -2914,7 +2914,7 @@ void Player::SendInitialSpells()
 
         data << uint32(itr->first);
 
-        data << uint16(itr->second.itemid);                 // cast item id
+        data << uint32(itr->second.itemid);                 // cast item id
         data << uint16(sEntry->GetCategory());              // spell category
 
         // send infinity cooldown in special format
